Move logging configs to api package
