+++ pkg/api/logging/config_env.go
@@ -1,10 +1,22 @@
-package slogutil
+package logging
 
 import (
 	"fmt"
 	"slices"
 	"strings"
+
+	"github.com/kelseyhightower/envconfig"
 )
+
+const (
+	LogHandlerConsole = "console"
+	LogHandlerOTLP    = "otlp"
+)
+
+var ALLOWED_LOG_HANDLERS = []string{
+	LogHandlerConsole,
+	LogHandlerOTLP,
+}
 
 type LogEnvironmentConfig struct {
 	Level        string      `envconfig:"LEVEL" default:"warn"`
@@ -12,6 +24,12 @@ type LogEnvironmentConfig struct {
 	ConsoleLevel string      `envconfig:"HANDLER_CONSOLE_LEVEL"`
 	OTLPLevel    string      `envconfig:"HANDLER_OTLP_LEVEL"`
 	OTLPEndpoint string      `envconfig:"HANDLER_OTLP_ENDPOINT"`
+}
+
+func LoadConfig() *LogEnvironmentConfig {
+	cfg := &LogEnvironmentConfig{}
+	_ = envconfig.Process("LOG", cfg)
+	return cfg
 }
 
 type LogHandlers []string
