--- pkg/lib/config/log_env.go
@@ -1,9 +1,0 @@
-package config
-
-import (
-	"github.com/authgear/authgear-server/pkg/util/slogutil"
-)
-
-type LogEnvironmentConfig struct {
-	slogutil.LogEnvironmentConfig
-}
